feat: introduce SponsorLogo component and factor out useScreenWidth hook
refs:
issue #20
