+++ components/sections/SplashScreen.tsx
@@ -10,6 +10,7 @@ import {
   TextStyledExtraBold,
 } from "@/components/StyledCore";
 import EmailListingInput from "@/components/EmailListingInput";
+import useScreenWidth from "../hooks/useScreenWidth";
 
 const minHeights = [550, 650, 700, 800];
 
@@ -20,24 +21,6 @@ const FlexFullView = styled(Flex)`
   background-size: 100vw auto;
   background-position: center bottom;
 `;
-
-function useScreenWidth() {
-  const [width, setWidth] = useState(0);
-
-  useEffect(() => {
-    const handleWidthChange = () => {
-      setWidth(window.innerWidth);
-    };
-    handleWidthChange();
-    window.addEventListener("resize", handleWidthChange);
-
-    return () => {
-      window.removeEventListener("resize", handleWidthChange);
-    };
-  }, []);
-
-  return width;
-}
 
 export default function SplashScreen(): React.ReactElement {
   const [heightProp, setHeightProp] = useState(
